fix: fix select styles
