+++ src/pages/network-visualisation.page.tsx
@@ -9,7 +9,7 @@ import { FilterControlsPanel } from '@/components/filter-controls-panel'
 import { MainContent } from '@/components/main-content'
 import { MultiComboBox } from '@/components/multi-combobox'
 import { RangeSlider } from '@/components/range-slider'
-import { Select } from '@/components/select'
+import { SingleSelect } from '@/components/single-select'
 import { db } from '@/db'
 import { type Gender, genders } from '@/db/genders'
 import { type EntityBase } from '@/db/types'
@@ -108,7 +108,7 @@ export default function GeoVisualisationPage(): JSX.Element {
                 onSelectionChange={filters.setSelectedProfessions}
                 selectedKeys={filters.selectedProfessions}
               />
-              <Select
+              <SingleSelect
                 items={labeledGenders}
                 label={t(['common', 'filter', 'gender'])}
                 messages={messages.gender}
